+++ test/typescript/global.ts
@@ -1,0 +1,13 @@
+import mongoose from 'mongoose';
+
+mongoose.set('useCreateIndex', true);
+
+mongoose.get('useCreateIndex');
+
+const m: mongoose.Mongoose = new mongoose.Mongoose();
+
+m.set('useUnifiedTopology', true);
+
+m.connect('mongodb://localhost:27017/test').then(() => {
+  console.log('Connected!');
+});